+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/test/java/io/flutter/plugins/googlemaps/ConvertTest.java
@@ -4,6 +4,7 @@
 
 package io.flutter.plugins.googlemaps;
 
+import static com.google.android.gms.maps.GoogleMap.MAP_TYPE_HYBRID;
 import static io.flutter.plugins.googlemaps.Convert.HEATMAP_DATA_KEY;
 import static io.flutter.plugins.googlemaps.Convert.HEATMAP_GRADIENT_COLORS_KEY;
 import static io.flutter.plugins.googlemaps.Convert.HEATMAP_GRADIENT_COLOR_MAP_SIZE_KEY;
@@ -16,7 +17,9 @@ import static io.flutter.plugins.googlemaps.Convert.HEATMAP_RADIUS_KEY;
 import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 
 import android.content.res.AssetManager;
@@ -29,6 +32,7 @@ import android.util.Base64;
 import androidx.annotation.NonNull;
 import com.google.android.gms.maps.model.BitmapDescriptor;
 import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.LatLngBounds;
 import com.google.maps.android.clustering.algo.StaticCluster;
 import com.google.maps.android.geometry.Point;
 import com.google.maps.android.heatmaps.Gradient;
@@ -64,10 +68,12 @@ public class ConvertTest {
 
   @Mock private FlutterInjectorWrapper flutterInjectorWrapper;
 
+  @Mock private GoogleMapOptionsSink optionsSink;
+
   AutoCloseable mockCloseable;
 
   // A 1x1 pixel (#8080ff) PNG image encoded in base64
-  private String base64Image = generateBase64Image();
+  private final String base64Image = generateBase64Image();
 
   @Before
   public void before() {
@@ -83,7 +89,7 @@ public class ConvertTest {
   public void ConvertToPointsConvertsThePointsWithFullPrecision() {
     double latitude = 43.03725568057;
     double longitude = -87.90466904649;
-    ArrayList<Double> point = new ArrayList<Double>();
+    ArrayList<Double> point = new ArrayList<>();
     point.add(latitude);
     point.add(longitude);
     ArrayList<ArrayList<Double>> pointsList = new ArrayList<>();
@@ -247,7 +253,7 @@ public class ConvertTest {
   }
 
   @Test
-  public void GetBitmapFromBytesAuto() throws Exception {
+  public void GetBitmapFromBytesAuto() {
     byte[] bmpData = Base64.decode(base64Image, Base64.DEFAULT);
 
     Map<String, Object> assetDetails = new HashMap<>();
@@ -264,7 +270,7 @@ public class ConvertTest {
   }
 
   @Test
-  public void GetBitmapFromBytesAutoAndWidth() throws Exception {
+  public void GetBitmapFromBytesAutoAndWidth() {
     byte[] bmpData = Base64.decode(base64Image, Base64.DEFAULT);
 
     Map<String, Object> assetDetails = new HashMap<>();
@@ -282,7 +288,7 @@ public class ConvertTest {
   }
 
   @Test
-  public void GetBitmapFromBytesAutoAndHeight() throws Exception {
+  public void GetBitmapFromBytesAutoAndHeight() {
     byte[] bmpData = Base64.decode(base64Image, Base64.DEFAULT);
 
     Map<String, Object> assetDetails = new HashMap<>();
@@ -300,7 +306,7 @@ public class ConvertTest {
   }
 
   @Test
-  public void GetBitmapFromBytesNoScaling() throws Exception {
+  public void GetBitmapFromBytesNoScaling() {
     byte[] bmpData = Base64.decode(base64Image, Base64.DEFAULT);
 
     Map<String, Object> assetDetails = new HashMap<>();
@@ -317,7 +323,7 @@ public class ConvertTest {
   }
 
   @Test(expected = IllegalArgumentException.class) // Expecting an IllegalArgumentException
-  public void GetBitmapFromBytesThrowsErrorIfInvalidImageData() throws Exception {
+  public void GetBitmapFromBytesThrowsErrorIfInvalidImageData() {
     String invalidBase64Image = "not valid image data";
     byte[] bmpData = Base64.decode(invalidBase64Image, Base64.DEFAULT);
 
@@ -336,6 +342,214 @@ public class ConvertTest {
     }
 
     fail("Expected an IllegalArgumentException to be thrown");
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesNulls() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verifyNoInteractions(optionsSink);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesCompassEnabled() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setCompassEnabled(false).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setCompassEnabled(false);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesMapToolbarEnabled() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setMapToolbarEnabled(true).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setMapToolbarEnabled(true);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesRotateGesturesEnabled() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setRotateGesturesEnabled(false).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setRotateGesturesEnabled(false);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesScrollGesturesEnabled() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setScrollGesturesEnabled(true).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setScrollGesturesEnabled(true);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesTiltGesturesEnabled() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setTiltGesturesEnabled(false).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setTiltGesturesEnabled(false);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesTrackCameraPosition() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setTrackCameraPosition(true).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setTrackCameraPosition(true);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesZoomControlsEnabled() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setZoomControlsEnabled(false).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setZoomControlsEnabled(false);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesZoomGesturesEnabled() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setZoomGesturesEnabled(true).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setZoomGesturesEnabled(true);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesMyLocationEnabled() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setMyLocationEnabled(false).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setMyLocationEnabled(false);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesMyLocationButtonEnabled() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setMyLocationButtonEnabled(true).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setMyLocationButtonEnabled(true);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesIndoorViewEnabled() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setIndoorViewEnabled(false).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setIndoorEnabled(false);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesTrafficEnabled() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setTrafficEnabled(true).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setTrafficEnabled(true);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesBuildingsEnabled() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setBuildingsEnabled(false).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setBuildingsEnabled(false);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesLiteModeEnabled() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setLiteModeEnabled(true).build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setLiteModeEnabled(true);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesStyle() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder().setStyle("foo").build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setMapStyle("foo");
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesUnboundedCameraTargetBounds() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder()
+            .setCameraTargetBounds(new Messages.PlatformCameraTargetBounds.Builder().build())
+            .build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setCameraTargetBounds(null);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesBoundedCameraTargetBounds() {
+    LatLngBounds bounds = new LatLngBounds(new LatLng(10, 20), new LatLng(30, 40));
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder()
+            .setCameraTargetBounds(
+                new Messages.PlatformCameraTargetBounds.Builder()
+                    .setBounds(
+                        new Messages.PlatformLatLngBounds.Builder()
+                            .setSouthwest(
+                                new Messages.PlatformLatLng.Builder()
+                                    .setLatitude(bounds.southwest.latitude)
+                                    .setLongitude(bounds.southwest.longitude)
+                                    .build())
+                            .setNortheast(
+                                new Messages.PlatformLatLng.Builder()
+                                    .setLatitude(bounds.northeast.latitude)
+                                    .setLongitude(bounds.northeast.longitude)
+                                    .build())
+                            .build())
+                    .build())
+            .build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setCameraTargetBounds(bounds);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesMapType() {
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder()
+            .setMapType(Messages.PlatformMapType.HYBRID)
+            .build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setMapType(MAP_TYPE_HYBRID);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesPadding() {
+    final double top = 1.0;
+    final double bottom = 2.0;
+    final double left = 3.0;
+    final double right = 4.0;
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder()
+            .setPadding(
+                new Messages.PlatformEdgeInsets.Builder()
+                    .setTop(top)
+                    .setBottom(bottom)
+                    .setLeft(left)
+                    .setRight(right)
+                    .build())
+            .build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1))
+        .setPadding((float) top, (float) left, (float) bottom, (float) right);
+  }
+
+  @Test
+  public void interpretMapConfiguration_handlesMinMaxZoomPreference() {
+    final double min = 1.0;
+    final double max = 2.0;
+    final Messages.PlatformMapConfiguration config =
+        new Messages.PlatformMapConfiguration.Builder()
+            .setMinMaxZoomPreference(
+                new Messages.PlatformZoomRange.Builder().setMin(min).setMax(max).build())
+            .build();
+    Convert.interpretMapConfiguration(config, optionsSink);
+    verify(optionsSink, times(1)).setMinMaxZoomPreference((float) min, (float) max);
   }
 
   private static final SphericalMercatorProjection sProjection = new SphericalMercatorProjection(1);
@@ -465,8 +679,7 @@ public class ConvertTest {
     ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
     fakeBitmap.compress(Bitmap.CompressFormat.PNG, 100, byteArrayOutputStream);
     byte[] byteArray = byteArrayOutputStream.toByteArray();
-    InputStream fakeStream = new ByteArrayInputStream(byteArray);
-    return fakeStream;
+    return new ByteArrayInputStream(byteArray);
   }
 
   // Helper method to generate 1x1 pixel base64 encoded png test image
@@ -487,9 +700,7 @@ public class ConvertTest {
     byte[] pngBytes = outputStream.toByteArray();
 
     // Encode the PNG bytes as a base64 string
-    String base64Image = Base64.encodeToString(pngBytes, Base64.DEFAULT);
-
-    return base64Image;
+    return Base64.encodeToString(pngBytes, Base64.DEFAULT);
   }
 }