+++ packages/google_maps_flutter/google_maps_flutter_ios/pigeons/messages.dart
@@ -12,6 +12,15 @@ import 'package:pigeon/pigeon.dart';
   copyrightHeader: 'pigeons/copyright.txt',
 ))
 
+/// Pigeon equivalent of MapType
+enum PlatformMapType {
+  none,
+  normal,
+  satellite,
+  terrain,
+  hybrid,
+}
+
 /// Pigeon representatation of a CameraPosition.
 class PlatformCameraPosition {
   PlatformCameraPosition({
@@ -64,6 +73,24 @@ class PlatformHeatmap {
   // TODO(stuartmorgan): Replace this with structured data. This exists only to
   //  allow incremental migration to Pigeon.
   final Object json;
+}
+
+/// Pigeon equivalent of Cluster.
+class PlatformCluster {
+  PlatformCluster({
+    required this.clusterManagerId,
+    required this.position,
+    required this.bounds,
+    required this.markerIds,
+  });
+
+  final String clusterManagerId;
+  final PlatformLatLng position;
+  final PlatformLatLngBounds bounds;
+  // TODO(stuartmorgan): Make the generic type non-nullable once supported.
+  // https://github.com/flutter/flutter/issues/97848
+  // The consuming code treats the entries as non-nullable.
+  final List<String?> markerIds;
 }
 
 /// Pigeon equivalent of the ClusterManager class.
@@ -130,6 +157,21 @@ class PlatformTileOverlay {
   final Object json;
 }
 
+/// Pigeon equivalent of Flutter's EdgeInsets.
+class PlatformEdgeInsets {
+  PlatformEdgeInsets({
+    required this.top,
+    required this.bottom,
+    required this.left,
+    required this.right,
+  });
+
+  final double top;
+  final double bottom;
+  final double left;
+  final double right;
+}
+
 /// Pigeon equivalent of LatLng.
 class PlatformLatLng {
   PlatformLatLng({required this.latitude, required this.longitude});
@@ -146,34 +188,83 @@ class PlatformLatLngBounds {
   final PlatformLatLng southwest;
 }
 
-/// Pigeon equivalent of Cluster.
-class PlatformCluster {
-  PlatformCluster({
-    required this.clusterManagerId,
-    required this.position,
-    required this.bounds,
-    required this.markerIds,
+/// Pigeon equivalent of CameraTargetBounds.
+///
+/// As with the Dart version, it exists to distinguish between not setting a
+/// a target, and having an explicitly unbounded target (null [bounds]).
+class PlatformCameraTargetBounds {
+  PlatformCameraTargetBounds({required this.bounds});
+
+  final PlatformLatLngBounds? bounds;
+}
+
+/// Information passed to the platform view creation.
+class PlatformMapViewCreationParams {
+  PlatformMapViewCreationParams({
+    required this.initialCameraPosition,
+    required this.mapConfiguration,
+    required this.initialCircles,
+    required this.initialMarkers,
+    required this.initialPolygons,
+    required this.initialPolylines,
+    required this.initialHeatmaps,
+    required this.initialTileOverlays,
+    required this.initialClusterManagers,
   });
 
-  final String clusterManagerId;
-  final PlatformLatLng position;
-  final PlatformLatLngBounds bounds;
-  // TODO(stuartmorgan): Make the generic type non-nullable once supported.
+  final PlatformCameraPosition initialCameraPosition;
+  final PlatformMapConfiguration mapConfiguration;
+  // TODO(stuartmorgan): Make the generic types non-nullable once supported.
   // https://github.com/flutter/flutter/issues/97848
   // The consuming code treats the entries as non-nullable.
-  final List<String?> markerIds;
+  final List<PlatformCircle?> initialCircles;
+  final List<PlatformMarker?> initialMarkers;
+  final List<PlatformPolygon?> initialPolygons;
+  final List<PlatformPolyline?> initialPolylines;
+  final List<PlatformHeatmap?> initialHeatmaps;
+  final List<PlatformTileOverlay?> initialTileOverlays;
+  final List<PlatformClusterManager?> initialClusterManagers;
 }
 
 /// Pigeon equivalent of MapConfiguration.
 class PlatformMapConfiguration {
-  PlatformMapConfiguration({required this.json});
+  PlatformMapConfiguration({
+    required this.compassEnabled,
+    required this.cameraTargetBounds,
+    required this.mapType,
+    required this.minMaxZoomPreference,
+    required this.rotateGesturesEnabled,
+    required this.scrollGesturesEnabled,
+    required this.tiltGesturesEnabled,
+    required this.trackCameraPosition,
+    required this.zoomGesturesEnabled,
+    required this.myLocationEnabled,
+    required this.myLocationButtonEnabled,
+    required this.padding,
+    required this.indoorViewEnabled,
+    required this.trafficEnabled,
+    required this.buildingsEnabled,
+    required this.cloudMapId,
+    required this.style,
+  });
 
-  /// The configuration options, as JSON. This should only be set from
-  /// _jsonForMapConfiguration, and the native code must interpret it according
-  /// to the internal implementation details of that method.
-  // TODO(stuartmorgan): Replace this with structured data. This exists only to
-  //  allow incremental migration to Pigeon.
-  final Object json;
+  final bool? compassEnabled;
+  final PlatformCameraTargetBounds? cameraTargetBounds;
+  final PlatformMapType? mapType;
+  final PlatformZoomRange? minMaxZoomPreference;
+  final bool? rotateGesturesEnabled;
+  final bool? scrollGesturesEnabled;
+  final bool? tiltGesturesEnabled;
+  final bool? trackCameraPosition;
+  final bool? zoomGesturesEnabled;
+  final bool? myLocationEnabled;
+  final bool? myLocationButtonEnabled;
+  final PlatformEdgeInsets? padding;
+  final bool? indoorViewEnabled;
+  final bool? trafficEnabled;
+  final bool? buildingsEnabled;
+  final String? cloudMapId;
+  final String? style;
 }
 
 /// Pigeon representation of an x,y coordinate.
@@ -203,8 +294,8 @@ class PlatformTileLayer {
 class PlatformZoomRange {
   PlatformZoomRange({required this.min, required this.max});
 
-  final double min;
-  final double max;
+  final double? min;
+  final double? max;
 }
 
 /// Interface for non-test interactions with the native SDK.
@@ -404,6 +495,15 @@ abstract class MapsCallbackApi {
   @ObjCSelector('tileWithOverlayIdentifier:location:zoom:')
   PlatformTile getTileOverlayTile(
       String tileOverlayId, PlatformPoint location, int zoom);
+}
+
+/// Dummy interface to force generation of the platform view creation params,
+/// which are not used in any Pigeon calls, only the platform view creation
+/// call made internally by Flutter.
+@HostApi()
+abstract class MapsPlatformViewApi {
+  // This is never actually called.
+  void createView(PlatformMapViewCreationParams? type);
 }
 
 /// Inspector API only intended for use in integration tests.